seq hopefully finished

+++ src/Point.java
@@ -9,7 +9,8 @@ public class Point {
     public double getY() {
         return y;
     }
-    public Point(int x, int y) {
+
+    public Point(double x, double y) {
         this.x = x;
         this.y = y;
     }
